✨ Create SplashFactory for dependency injection and modularization
